Fix SVACE issue
DO_NOT_ASSIGN_SIZE_OF_INT
DO_NOT_CAST_LONG_TO_DOUBLE
INVALID_ASSIGNMENT_WITH_UNSIGNED_VALUES

+++ dali/internal/adaptor/common/combined-update-render-controller.cpp
@@ -593,8 +593,8 @@ void CombinedUpdateRenderController::UpdateRenderThread()
     // UPDATE
     //////////////////////////////
 
-    const unsigned int currentTime   = currentFrameStartTime / NANOSECONDS_PER_MILLISECOND;
-    const unsigned int nextFrameTime = currentTime + mDefaultFrameDurationMilliseconds;
+    const uint32_t currentTime   = static_cast<uint32_t>(currentFrameStartTime / NANOSECONDS_PER_MILLISECOND);
+    const uint32_t nextFrameTime = currentTime + static_cast<uint32_t>(mDefaultFrameDurationMilliseconds);
 
     uint64_t noOfFramesSinceLastUpdate = 1;
     float    frameDelta                = 0.0f;
